# packages/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BookWorm API
  version: 1.0.0
  description: |
    REST API for **BookWorm**, an e-bookstore where guests and registered users browse, buy and track books.

    **Conventions**
    - Money is stored as integer **paise**. Every money field is returned twice: `xxxPaise` (integer) and `xxxInr` (formatted, e.g. `₹149`).
    - Errors always look like `{ "error": { "code", "message", "details" } }`.
    - Paginated lists look like `{ "items", "page", "pageSize", "total", "totalPages" }`.
    - Authenticate with `Authorization: Bearer <token>` from `/auth/login`, `/auth/register` or `/auth/guest-session`.
    - Guest tokens (role `GUEST`) only work for cart, checkout and payment of orders placed in the same guest session.

    Demo accounts: `customer@test.com / Test@1234`, `admin@bookworm.com / Admin@1234`.
  contact:
    name: BookWorm Team
    email: admin@bookworm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: /api
    description: Same origin (Docker / nginx)
security:
  - BearerAuth: []
...
paths:
...
  /auth/guest-session:
    post:
      tags: [Auth]
      operationId: createGuestSession
      summary: Start a guest checkout session with just an email
      description: |
        Returns a 24h guest token. Fails with `409 ACCOUNT_EXISTS` if the email belongs to a registered account
        (the user must log in instead). Each call starts a new guest session (`gsid`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GuestSessionRequest' }
      responses:
        '200':
          description: Guest session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/profile:
    get:
      tags: [Auth]
      operationId: getProfile
      summary: Get the current user's profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [Auth]
      operationId: updateProfile
      summary: Update name or phone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/set-password:
    put:
      tags: [Auth]
      operationId: setPassword
      summary: Convert a guest into a registered customer
      description: |
        Guest token required. Only allowed when the guest placed at least one order in the current guest session.
        Keeps the same user id, so all previous orders stay visible in My Orders.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetPasswordRequest' }
      responses:
        '200':
          description: Account converted; new 7-day token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ───────────────────────────── Users ─────────────────────────────
...
  /orders/checkout:
    post:
      tags: [Orders]
      operationId: checkout
      summary: Create a PENDING order from my cart and reserve stock for 30 minutes
      description: |
        Provide either `addressId` (saved address) or `address` (typed in). The cart is not cleared and the coupon
        and gift points are not consumed until payment succeeds.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutRequest' }
      responses:
        '201':
          description: Order created and stock reserved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: My order history (newest first)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Page of orders
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /orders/lookup:
    post:
      tags: [Orders]
      operationId: lookupOrder
      summary: Track an order with email + order number (no login)
      description: Rate limited per IP. Returns a generic 404 when the email and order number don't match.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderLookupRequest' }
      responses:
        '200':
          description: Order status and tracking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackedOrder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Order details (owner only; guests only for their current session)
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orders/{orderId}/cancel:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags: [Orders]
      operationId: cancelOrder
      summary: Cancel within 48 hours (before shipping) — restores stock and refunds
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /orders/{orderId}/return:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags: [Orders]
      operationId: returnOrder
      summary: Request a return within 7 days of delivery — creates a return shipment
      responses:
        '200':
          description: Return requested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /orders/{orderId}/address:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags: [Orders]
      operationId: updateOrderAddress
      summary: Change the delivery address before the order ships
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ───────────────────────────── Payments ─────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 48, default: 12 }
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: c4690a4e-c706-59e0-aad9-b3190af2b98e
...
  responses:
    BadRequest:
      description: The request is invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: 'request/body/email must match format "email"'
              details: [{ path: /body/email, message: 'must match format "email"' }]
    Unauthorized:
      description: Missing, invalid or expired token, or wrong credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: UNAUTHORIZED, message: Authentication required }
    Forbidden:
      description: Authenticated but not allowed (e.g. guest or non-admin)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: FORBIDDEN, message: You do not have access to this resource }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: NOT_FOUND, message: Book not found }
    Conflict:
      description: Conflicts with the current state (duplicate, out of stock, wrong status, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: INSUFFICIENT_STOCK, message: Only 2 copies of Joy of Minimalism are available }
    Unprocessable:
      description: Business rule failed (e.g. invalid coupon or not enough gift points)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: COUPON_EXPIRED, message: Coupon EXPIRED10 has expired }
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: RATE_LIMITED, message: 'Too many requests, please try again later' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: NOT_FOUND }
            message: { type: string, example: Book not found }
            details: {}
    Uuid:
      type: string
      format: uuid
...
    Paise:
      type: integer
      minimum: 0
      description: Amount in paise (1 INR = 100 paise)
    Inr:
      type: string
      description: Formatted INR amount
      example: ₹149
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: integer, minimum: 1, example: 1 }
        pageSize: { type: integer, minimum: 1, example: 12 }
        total: { type: integer, minimum: 0, example: 36 }
        totalPages: { type: integer, minimum: 0, example: 3 }
...
    UserRole: { type: string, enum: [GUEST, CUSTOMER, ADMIN] }
    BookFormat: { type: string, enum: [PAPERBACK, HARDCOVER, EBOOK] }
    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED, EXPIRED, RETURN_REQUESTED, RETURNED]
    PaymentStatus: { type: string, enum: [UNPAID, PAID, FAILED, REFUNDED] }
    PaymentMethod: { type: string, enum: [CREDIT_CARD, DEBIT_CARD, UPI, WALLET] }
    PaymentTransactionStatus: { type: string, enum: [INITIATED, SUCCEEDED, FAILED, REFUNDED] }
    ShipmentType: { type: string, enum: [FORWARD, RETURN] }
    ShipmentStatus: { type: string, enum: [PROCESSING, SHIPPED, OUT_FOR_DELIVERY, DELIVERED] }
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, giftPoints, walletBalancePaise, walletBalanceInr]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/UserRole' }
        giftPoints: { type: integer, minimum: 0 }
        walletBalancePaise: { $ref: '#/components/schemas/Paise' }
        walletBalanceInr: { $ref: '#/components/schemas/Inr' }
        createdAt: { type: string, format: date-time }
      example:
        id: 73165c5c-dfb7-53cc-ab79-777827a0a909
        email: customer@test.com
        firstName: John
        lastName: Smith
        phone: '9876543210'
        role: CUSTOMER
        giftPoints: 200
        walletBalancePaise: 100000
        walletBalanceInr: ₹1,000
        createdAt: '2026-10-04T10:00:00.000Z'
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string, description: JWT, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
        user: { $ref: '#/components/schemas/User' }
    Password:
      type: string
      minLength: 8
      maxLength: 72
      pattern: '^(?=.*[A-Z])(?=.*\d).+$'
      description: At least 8 characters with one uppercase letter and one number
      example: Test@1234
...
    GuestSessionRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
      example: { email: guest.reader@example.com }
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        firstName: { type: string, minLength: 1, maxLength: 50 }
        lastName: { type: string, minLength: 1, maxLength: 50 }
        phone: { type: string, pattern: '^\d{10}$' }
      example: { firstName: John, lastName: Smith }
    SetPasswordRequest:
      type: object
      additionalProperties: false
      required: [password, confirmPassword]
      properties:
        password: { $ref: '#/components/schemas/Password' }
        confirmPassword: { type: string, maxLength: 72 }
      example: { password: Reader@123, confirmPassword: Reader@123 }
    AddressInput:
      type: object
      additionalProperties: false
      required: [firstName, lastName, email, phone, line1, city, pin, state]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 50 }
        lastName: { type: string, minLength: 1, maxLength: 50 }
        email: { type: string, format: email, maxLength: 254 }
        phone: { type: string, pattern: '^\d{10}$', description: 10-digit Indian mobile number (without +91) }
        line1: { type: string, minLength: 1, maxLength: 120 }
        line2: { type: string, maxLength: 120 }
        city: { type: string, minLength: 1, maxLength: 60 }
        pin: { type: string, pattern: '^\d{6}$' }
        state: { type: string, minLength: 1, maxLength: 60 }
        country: { type: string, maxLength: 60, default: India }
      example:
        firstName: John
        lastName: Smith
        email: customer@test.com
        phone: '9876543210'
        line1: 221 MG Road
        line2: Near Trinity Metro Station
        city: Bengaluru
        pin: '560001'
        state: Karnataka
        country: India
    ShippingAddress:
      type: object
      required: [firstName, lastName, email, phone, line1, city, pin, state, country]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        pin: { type: string }
        state: { type: string }
        country: { type: string }
...
    OrderTotals:
      type: object
      required:
        - itemCount
        - subtotalPaise
        - subtotalInr
        - taxPaise
        - taxInr
        - deliveryChargePaise
        - deliveryChargeInr
        - couponDiscountPaise
        - couponDiscountInr
        - giftDiscountPaise
        - giftDiscountInr
        - totalPaise
        - totalInr
        - giftPointsRedeemed
        - giftPointsEarned
      properties:
        itemCount: { type: integer, minimum: 0 }
        subtotalPaise: { $ref: '#/components/schemas/Paise' }
        subtotalInr: { $ref: '#/components/schemas/Inr' }
        taxPaise: { $ref: '#/components/schemas/Paise' }
        taxInr: { $ref: '#/components/schemas/Inr' }
        deliveryChargePaise: { $ref: '#/components/schemas/Paise' }
        deliveryChargeInr: { $ref: '#/components/schemas/Inr' }
        couponDiscountPaise: { $ref: '#/components/schemas/Paise' }
        couponDiscountInr: { $ref: '#/components/schemas/Inr' }
        giftDiscountPaise: { $ref: '#/components/schemas/Paise' }
        giftDiscountInr: { $ref: '#/components/schemas/Inr' }
        totalPaise: { $ref: '#/components/schemas/Paise' }
        totalInr: { $ref: '#/components/schemas/Inr' }
        giftPointsRedeemed: { type: integer, minimum: 0 }
        giftPointsEarned: { type: integer, minimum: 0 }
      example:
        itemCount: 2
        subtotalPaise: 59800
        subtotalInr: ₹598
        taxPaise: 7176
        taxInr: ₹71.76
        deliveryChargePaise: 0
        deliveryChargeInr: ₹0
        couponDiscountPaise: 5000
        couponDiscountInr: ₹50
        giftDiscountPaise: 0
        giftDiscountInr: ₹0
        totalPaise: 61976
        totalInr: ₹619.76
        giftPointsRedeemed: 0
        giftPointsEarned: 30
    OrderFlags:
      type: object
      required: [canCancel, canReturn, canModifyAddress]
      properties:
        canCancel: { type: boolean }
        canReturn: { type: boolean }
        canModifyAddress: { type: boolean }
    OrderItem:
      type: object
      required: [id, bookId, title, quantity, priceAtPurchasePaise, priceAtPurchaseInr, lineTotalPaise, lineTotalInr]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        bookId: { $ref: '#/components/schemas/Uuid' }
        title: { type: string }
        authorName: { type: string }
        coverImageUrl: { type: string }
        format: { $ref: '#/components/schemas/BookFormat' }
        quantity: { type: integer, minimum: 1 }
        priceAtPurchasePaise: { $ref: '#/components/schemas/Paise' }
        priceAtPurchaseInr: { $ref: '#/components/schemas/Inr' }
        lineTotalPaise: { $ref: '#/components/schemas/Paise' }
        lineTotalInr: { $ref: '#/components/schemas/Inr' }
        deliveryDate: { type: string, format: date-time, nullable: true }
    OrderSummary:
      type: object
      required: [id, orderNumber, status, paymentStatus, createdAt, totals, items, flags]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        orderNumber: { type: string, pattern: '^BW-[0-9A-Z]{8}$', example: BW-SEED000B }
        status: { $ref: '#/components/schemas/OrderStatus' }
        paymentStatus: { $ref: '#/components/schemas/PaymentStatus' }
        createdAt: { type: string, format: date-time }
        totals: { $ref: '#/components/schemas/OrderTotals' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        flags: { $ref: '#/components/schemas/OrderFlags' }
        shipments:
          type: array
          description: Included in the admin order list so shipments can be advanced from there.
          items: { $ref: '#/components/schemas/Shipment' }
    OrderPage:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/OrderSummary' }
    Order:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          required: [contactEmail, shippingAddress, payments, shipments]
          properties:
            contactEmail: { type: string, format: email }
            shippingAddress: { $ref: '#/components/schemas/ShippingAddress' }
            paymentMethod:
              type: string
              nullable: true
              enum: [CREDIT_CARD, DEBIT_CARD, UPI, WALLET, null]
            couponCode: { type: string, nullable: true }
            reservedUntil: { type: string, format: date-time, nullable: true }
            confirmedAt: { type: string, format: date-time, nullable: true }
            cancelledAt: { type: string, format: date-time, nullable: true }
            payments:
              type: array
              items: { $ref: '#/components/schemas/Payment' }
            shipments:
              type: array
              items: { $ref: '#/components/schemas/Shipment' }
    OrderResponse:
      type: object
      required: [order]
      properties:
        order: { $ref: '#/components/schemas/Order' }
    CheckoutRequest:
      type: object
      additionalProperties: false
      required: [paymentMethod]
      description: Provide exactly one of `addressId` or `address`.
      properties:
        addressId: { $ref: '#/components/schemas/Uuid' }
        address: { $ref: '#/components/schemas/AddressInput' }
        saveAddress: { type: boolean, default: false, description: Registered users only }
        couponCode: { type: string, minLength: 1, maxLength: 30 }
        giftPointsToRedeem: { type: integer, minimum: 0, default: 0 }
        paymentMethod: { $ref: '#/components/schemas/PaymentMethod' }
      example:
        addressId: 8e8e8f42-3d3f-5745-a311-5df967b495cf
        couponCode: BOOK10
        giftPointsToRedeem: 200
        paymentMethod: CREDIT_CARD
    OrderLookupRequest:
      type: object
      additionalProperties: false
      required: [email, orderNumber]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        orderNumber: { type: string, pattern: '^[Bb][Ww]-[0-9A-Za-z]{8}$' }
      example: { email: customer@test.com, orderNumber: BW-SEED000B }
    TrackedOrder:
      type: object
      description: Public view of an order for the Track Order page (no address or payment details).
      required: [orderNumber, status, paymentStatus, createdAt, items, totals, shipments]
      properties:
        orderNumber: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        paymentStatus: { $ref: '#/components/schemas/PaymentStatus' }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        totals: { $ref: '#/components/schemas/OrderTotals' }
        shipments:
          type: array
          items: { $ref: '#/components/schemas/Shipment' }

    # ── Payments ──
    Payment:
      type: object
      required: [id, sessionId, method, amountPaise, amountInr, status, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        sessionId: { type: string }
        method: { $ref: '#/components/schemas/PaymentMethod' }
        amountPaise: { $ref: '#/components/schemas/Paise' }
        amountInr: { $ref: '#/components/schemas/Inr' }
        status: { $ref: '#/components/schemas/PaymentTransactionStatus' }
        cardLast4: { type: string, nullable: true, pattern: '^\d{4}$' }
        upiHandleMasked: { type: string, nullable: true }
        failureReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        refundedAt: { type: string, format: date-time, nullable: true }
...
    ShipmentEvent:
      type: object
      required: [status, note, occurredAt]
      properties:
        status: { $ref: '#/components/schemas/ShipmentStatus' }
        note: { type: string }
        occurredAt: { type: string, format: date-time }
    Shipment:
      type: object
      required: [id, type, trackingNumber, carrier, status, estimatedDelivery, estimatedDeliveryText, events]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        type: { $ref: '#/components/schemas/ShipmentType' }
        trackingNumber: { type: string, example: TRK-SEED00000B }
        carrier: { type: string, example: BookWorm Express }
        status: { $ref: '#/components/schemas/ShipmentStatus' }
        estimatedDelivery: { type: string, format: date-time }
        estimatedDeliveryText: { type: string, example: 'Delivery by Thu, 1 Oct' }
        actualDelivery: { type: string, format: date-time, nullable: true }
        shippingRatePaise: { $ref: '#/components/schemas/Paise' }
        shippingRateInr: { $ref: '#/components/schemas/Inr' }
        events:
          type: array
          items: { $ref: '#/components/schemas/ShipmentEvent' }
...